+++ src/TestSuite/TestCase.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace MeCms\TestSuite;
 
 use Cake\Cache\Cache;
+use Cake\Datasource\ConnectionManager;
 use MeTools\TestSuite\TestCase as BaseTestCase;
 
 /**
@@ -55,5 +56,15 @@ abstract class TestCase extends BaseTestCase
             Cache::getConfig($cacheKey) ?: $this->fail('Cache key `' . $cacheKey . '` does not exist');
             Cache::clear($cacheKey);
         }
+    }
+
+    /**
+     * Checks if the current `test` scheme is `mysql`
+     * @return bool
+     * @since 2.30.3
+     */
+    protected function isMySql(): bool
+    {
+        return ConnectionManager::get('test')->config()['scheme'] == 'mysql';
     }
 }