+++ tests/TestCase/Model/Table/PostsTableTest.php
@@ -138,9 +138,10 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
     public function testFindMethods(): void
     {
         $query = $this->Table->find('forIndex');
-        $sql = $query->sql();
         $this->assertArrayKeysEqual(['Categories', 'Tags', 'Users'], $query->getContain());
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_categories` `Categories` ON `Categories`.`id` = (`Posts`.`category_id`) INNER JOIN `users` `Users` ON `Users`.`id` = (`Posts`.`user_id`) ORDER BY `Posts`.`created` DESC', $sql);
+
+        $this->skipIf(!$this->isMySql());
+        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_categories` `Categories` ON `Categories`.`id` = (`Posts`.`category_id`) INNER JOIN `users` `Users` ON `Users`.`id` = (`Posts`.`user_id`) ORDER BY `Posts`.`created` DESC', $query->sql());
     }
 
     /**
@@ -198,6 +199,7 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
      */
     public function testQueryFromFilter(): void
     {
+        $this->skipIf(!$this->isMySql());
         $query = $this->Table->queryFromFilter($this->Table->find(), ['tag' => 'test']);
         $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = (`PostsTags`.`post_id`) INNER JOIN `tags` `Tags` ON (`tag` = :c0 AND `Tags`.`id` = (`PostsTags`.`tag_id`))', $query->sql());
         $this->assertEquals('test', $query->getValueBinder()->bindings()[':c0']['value']);
@@ -209,6 +211,7 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
      */
     public function testQueryForRelated(): void
     {
+        $this->skipIf(!$this->isMySql());
         $this->loadFixtures();
 
         $query = $this->Table->queryForRelated(4, true);